+++ src/Docker.Registry.DotNet/Authentication/AuthenticationProvider.cs
@@ -3,6 +3,8 @@ using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Threading.Tasks;
 
+using Docker.Registry.DotNet.Helpers;
+
 namespace Docker.Registry.DotNet.Authentication
 {
     /// <summary>
