+++ src/Docker.Registry.DotNet/Endpoints/BlobOperations.cs
@@ -2,7 +2,10 @@
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
+
+using Docker.Registry.DotNet.Helpers;
 using Docker.Registry.DotNet.Models;
+using Docker.Registry.DotNet.Registry;
 
 namespace Docker.Registry.DotNet.Endpoints
 {
