+++ src/Docker.Registry.DotNet/Helpers/EnumerableExtensions.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+
+namespace Docker.Registry.DotNet.Helpers
+{
+    public static class EnumerableExtensions
+    {
+        public static IEnumerable<T> IfNullEmpty<T>(this IEnumerable<T> enumerable)
+        {
+            return enumerable ?? Enumerable.Empty<T>();
+        }
+    }
+}
