+++ src/Docker.Registry.DotNet/Helpers/IQueryString.cs
@@ -1,0 +1,7 @@
+﻿namespace Docker.Registry.DotNet.Helpers
+{
+    internal interface IQueryString
+    {
+        string GetQueryString();
+    }
+}
