--- src/Docker.Registry.DotNet/IQueryString.cs
@@ -1,7 +1,0 @@
-﻿namespace Docker.Registry.DotNet
-{
-    internal interface IQueryString
-    {
-        string GetQueryString();
-    }
-}
