+++ src/Docker.Registry.DotNet/Registry/UnauthorizedApiException.cs
@@ -1,0 +1,12 @@
+﻿namespace Docker.Registry.DotNet.Registry
+{
+    /// <summary>
+    /// Thrown when an api response is returned as unauthorized.
+    /// </summary>
+    public class UnauthorizedApiException : RegistryApiException
+    {
+        internal UnauthorizedApiException(RegistryApiResponse response) : base(response)
+        {
+        }
+    }
+}
