--- src/Docker.Registry.DotNet/UnauthorizedApiException.cs
@@ -1,12 +1,0 @@
-﻿namespace Docker.Registry.DotNet
-{
-    /// <summary>
-    /// Thrown when an api response is returned as unauthorized.
-    /// </summary>
-    public class UnauthorizedApiException : RegistryApiException
-    {
-        internal UnauthorizedApiException(RegistryApiResponse response) : base(response)
-        {
-        }
-    }
-}
